refacor(ui): reorganize code tree
- Update copyright
