+++ app/src/main/java/com/osfans/trime/data/prefs/PreferenceDelegateUi.kt
@@ -10,7 +10,7 @@ import androidx.preference.EditTextPreference
 import androidx.preference.ListPreference
 import androidx.preference.Preference
 import androidx.preference.SwitchPreference
-import com.osfans.trime.ui.components.DialogSeekBarPreference
+import com.osfans.trime.ui.main.settings.DialogSeekBarPreference
 import com.osfans.trime.ui.main.settings.EditTextIntPreference
 
 abstract class PreferenceDelegateUi<T : Preference>(
