+++ app/src/main/java/com/osfans/trime/ui/common/OnItemChangedListener.kt
@@ -1,8 +1,8 @@
 /*
- * SPDX-License-Identifier: LGPL-2.1-or-later
- * SPDX-FileCopyrightText: Copyright 2021-2023 Fcitx5 for Android Contributors
+ * SPDX-FileCopyrightText: 2015 - 2025 Rime community
+ * SPDX-License-Identifier: GPL-3.0-or-later
  */
-package com.osfans.trime.ui.components
+package com.osfans.trime.ui.common
 
 /**
  * Functions are called after the container changed
