+++ app/src/main/java/com/osfans/trime/ui/common/ProgressBarDialogIndeterminate.kt
@@ -1,9 +1,9 @@
 /*
- * SPDX-FileCopyrightText: 2015 - 2024 Rime community
+ * SPDX-FileCopyrightText: 2015 - 2025 Rime community
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
 
-package com.osfans.trime.ui.components
+package com.osfans.trime.ui.common
 
 import android.content.Context
 import androidx.annotation.StringRes
