+++ app/src/main/java/com/osfans/trime/ui/main/MainFragment.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
 
-package com.osfans.trime.ui.fragments
+package com.osfans.trime.ui.main
 
 import android.os.Bundle
 import androidx.fragment.app.activityViewModels
@@ -11,8 +11,7 @@ import androidx.navigation.fragment.findNavController
 import androidx.preference.Preference
 import androidx.preference.get
 import com.osfans.trime.R
-import com.osfans.trime.ui.components.PaddingPreferenceFragment
-import com.osfans.trime.ui.main.MainViewModel
+import com.osfans.trime.ui.common.PaddingPreferenceFragment
 import splitties.resources.styledColor
 
 class MainFragment : PaddingPreferenceFragment() {
