+++ app/src/main/java/com/osfans/trime/ui/main/ToolkitFragment.kt
@@ -1,16 +1,16 @@
-// SPDX-FileCopyrightText: 2015 - 2024 Rime community
-//
-// SPDX-License-Identifier: GPL-3.0-or-later
+/*
+ * SPDX-FileCopyrightText: 2015 - 2025 Rime community
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
 
-package com.osfans.trime.ui.fragments
+package com.osfans.trime.ui.main
 
 import android.app.AlertDialog
 import android.os.Bundle
 import androidx.fragment.app.activityViewModels
 import androidx.preference.Preference
 import com.osfans.trime.R
-import com.osfans.trime.ui.components.PaddingPreferenceFragment
-import com.osfans.trime.ui.main.MainViewModel
+import com.osfans.trime.ui.common.PaddingPreferenceFragment
 import com.osfans.trime.util.AppUtils
 import com.osfans.trime.util.Logcat
 
@@ -42,7 +42,7 @@ class ToolkitFragment : PaddingPreferenceFragment() {
                         .Builder(context)
                         .setMessage(R.string.real_time_logs_confirm)
                         .setPositiveButton(R.string.ok) { _, _ ->
-                            Logcat.default.clearLog()
+                            Logcat.Companion.default.clearLog()
                         }.setNegativeButton(R.string.cancel, null)
                         .show()
                     true
