+++ app/src/main/java/com/osfans/trime/ui/main/log/LogView.kt
@@ -1,8 +1,9 @@
-// SPDX-FileCopyrightText: 2015 - 2024 Rime community
-//
-// SPDX-License-Identifier: GPL-3.0-or-later
+/*
+ * SPDX-FileCopyrightText: 2015 - 2025 Rime community
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
 
-package com.osfans.trime.ui.components.log
+package com.osfans.trime.ui.main.log
 
 import android.content.Context
 import android.util.AttributeSet
