+++ app/src/main/java/com/osfans/trime/ui/main/settings/FolderPickerPreference.kt
@@ -1,8 +1,9 @@
-// SPDX-FileCopyrightText: 2015 - 2024 Rime community
-//
-// SPDX-License-Identifier: GPL-3.0-or-later
+/*
+ * SPDX-FileCopyrightText: 2015 - 2025 Rime community
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
 
-package com.osfans.trime.ui.components
+package com.osfans.trime.ui.main.settings
 
 import android.content.Context
 import android.content.res.TypedArray
