+++ app/src/main/res/xml/keyboard_preference.xml
@@ -41,7 +41,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
             app:iconSpaceReserved="false"
             app:useSimpleSummaryProvider="true" />
 
-        <com.osfans.trime.ui.components.DialogSeekBarPreference
+        <com.osfans.trime.ui.main.settings.DialogSeekBarPreference
             android:defaultValue="80"
             android:key="keyboard__split_space"
             android:title="@string/keyboard__split_space_title"
@@ -135,7 +135,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
             android:dependency="custom_sound_effect_enabled"
             app:iconSpaceReserved="false"
             app:useSimpleSummaryProvider="true" />
-        <com.osfans.trime.ui.components.DialogSeekBarPreference
+        <com.osfans.trime.ui.main.settings.DialogSeekBarPreference
             android:defaultValue="0"
             android:dependency="sound_on_keypress"
             android:key="sound_volume"
@@ -161,7 +161,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
             android:title="@string/vibrate_on_key_repeat"
             android:dependency="vibrate_on_key_press"
             app:iconSpaceReserved="false" />
-        <com.osfans.trime.ui.components.DialogSeekBarPreference
+        <com.osfans.trime.ui.main.settings.DialogSeekBarPreference
             android:defaultValue="0"
             android:dependency="vibrate_on_key_press"
             android:key="vibration_duration"
@@ -173,7 +173,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
             app:systemDefaultValueText="@string/system_default"
             app:unit="ms"
             app:useSimpleSummaryProvider="true" />
-        <com.osfans.trime.ui.components.DialogSeekBarPreference
+        <com.osfans.trime.ui.main.settings.DialogSeekBarPreference
             android:defaultValue="0"
             android:dependency="vibrate_on_key_press"
             android:key="vibration_amplitude"
@@ -197,7 +197,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
             android:title="@string/keyboard__swipe_enabled"
             android:defaultValue="true"
             app:iconSpaceReserved="false" />
-        <com.osfans.trime.ui.components.DialogSeekBarPreference
+        <com.osfans.trime.ui.main.settings.DialogSeekBarPreference
             android:defaultValue="30"
             android:key="keyboard__key_swipe_travel"
             android:title="@string/keyboard__key_swipe_travel_title"
@@ -207,7 +207,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
             app:seekBarIncrement="10"
             app:useSimpleSummaryProvider="true"
             app:unit="dp" />
-        <com.osfans.trime.ui.components.DialogSeekBarPreference
+        <com.osfans.trime.ui.main.settings.DialogSeekBarPreference
             android:defaultValue="800"
             android:key="keyboard__key_swipe_velocity"
             android:title="@string/keyboard__key_swipe_velocity"
@@ -217,7 +217,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
             app:seekBarIncrement="100"
             app:useSimpleSummaryProvider="true"
             app:unit="dp/s"/>
-        <com.osfans.trime.ui.components.DialogSeekBarPreference
+        <com.osfans.trime.ui.main.settings.DialogSeekBarPreference
             android:defaultValue="300"
             android:key="keyboard__key_long_press_timeout"
             android:title="@string/keyboard__long_press_timeout_title"
@@ -229,7 +229,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
             app:seekBarIncrement="10"
             app:unit="ms"
             app:useSimpleSummaryProvider="true" />
-        <com.osfans.trime.ui.components.DialogSeekBarPreference
+        <com.osfans.trime.ui.main.settings.DialogSeekBarPreference
             android:defaultValue="30"
             android:key="keyboard__key_repeat_interval"
             android:title="@string/keyboard__key_repeat_interval_title"
@@ -240,7 +240,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
             app:seekBarIncrement="10"
             app:unit="ms"
             app:useSimpleSummaryProvider="true" />
-        <com.osfans.trime.ui.components.DialogSeekBarPreference
+        <com.osfans.trime.ui.main.settings.DialogSeekBarPreference
             android:defaultValue="300"
             android:key="keyboard__key_double_tap_timeout"
             android:title="@string/keyboard__key_double_tap_timeout"
@@ -251,7 +251,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
             app:seekBarIncrement="10"
             app:unit="ms"
             app:useSimpleSummaryProvider="true" />
-        <com.osfans.trime.ui.components.DialogSeekBarPreference
+        <com.osfans.trime.ui.main.settings.DialogSeekBarPreference
             android:defaultValue="24"
             android:key="keyboard__key_slide_step_size"
             android:title="@string/keyboard__key_slide_step_size"
